refactor: ♻️ search 페이지 내부 컴포넌트 위치 이동
