+++ src/containers/SearchContainer/SearchItem/SearchItem.stories.tsx
@@ -3,7 +3,7 @@ import { ComponentStory, ComponentMeta } from '@storybook/react';
 import SearchItem from './SearchItem';
 
 export default {
-  title: 'Components/SearchItem',
+  title: 'Components/search/SearchItem',
   component: SearchItem,
   argTypes: {
     type: { control: 'radio', options: ['search', 'history'] },
